fix(backend): complete Decimal/float alignment and redis deprecation
- models/refund.py: change refund_amount from float to Decimal to match
  Order.total_amount and the Numeric SQLAlchemy column.
- test_refund_service.py, test_refund_tasks.py, test_logistics_tool.py,
  test_payment_tool.py, test_admin_api.py, test_admin_service.py:
  update all float literals to Decimal for total_amount and refund_amount.
- api/v1/admin/agent_config.py: replace deprecated redis.close() with
  redis.aclose().

+++ app/models/refund.py
@@ -1,5 +1,6 @@
 # app/models/refund.py
 from datetime import datetime
+from decimal import Decimal
 from enum import Enum
 
 from sqlalchemy import Column, DateTime, Numeric, String, Text, text
@@ -60,7 +61,7 @@ class RefundApplication(SQLModel, table=True):
     )
 
     # 退款金额（可能部分退款）
-    refund_amount: float = Field(
+    refund_amount: Decimal = Field(
         sa_column=Column(Numeric(precision=10, scale=2)), description="退款金额"
     )
